--- src/app/principal/page.tsx
@@ -1,9 +1,0 @@
-import Principal from "@/modules/auth/templates/principal";
-
-export const metadata = {
-  title: "OptiTech",
-};
-
-export default function Page() {
-  return <Principal />;
-}
