+++ src/auth.config.ts
@@ -1,12 +1,16 @@
 import { NextAuthConfig } from "next-auth";
 import { NextResponse } from "next/server";
 import { ROUTES_AUTH } from "./modules/auth/types/auth";
+import path from "path";
 
 const publicRoutes = new Set<string>([
     ROUTES_AUTH.LOGIN,
     ROUTES_AUTH.SING_UP,
-    ROUTES_AUTH.PRINCIPAL,
     ROUTES_AUTH.RESET_PASSWORD
+]);
+
+const secureRoutes = new Set<string>([
+    ROUTES_AUTH.DASHBOARD
 ]);
 
 export const authConfig: NextAuthConfig = {
@@ -19,8 +23,12 @@ export const authConfig: NextAuthConfig = {
             const isLoggedIn = !!auth?.user;
             const { pathname } = nextUrl;
 
-            if (isLoggedIn && pathname === ROUTES_AUTH.LOGIN) {
+            if(!isLoggedIn && secureRoutes.has(pathname)){
                 return NextResponse.redirect(new URL('/', nextUrl));
+            }
+
+            if (isLoggedIn && pathname === ROUTES_AUTH.LOGIN) {
+                return NextResponse.redirect(new URL(ROUTES_AUTH.DASHBOARD, nextUrl));
             }
 
             if (!isLoggedIn && publicRoutes.has(pathname)) {
@@ -28,17 +36,13 @@ export const authConfig: NextAuthConfig = {
             }
 
             if (isLoggedIn && publicRoutes.has(pathname)) {
-                return NextResponse.redirect(new URL('/', nextUrl));
+                return NextResponse.redirect(new URL(ROUTES_AUTH.DASHBOARD, nextUrl));
             }
-
-            if (!isLoggedIn) {
-                return NextResponse.redirect(new URL(ROUTES_AUTH.PRINCIPAL, nextUrl));
-            }
-
+            
             return true;
         },
     },
     debug: process.env.NODE_ENV !== "production",
-    basePath: '/',
+    basePath: ROUTES_AUTH.DASHBOARD,
     providers: [],
 } satisfies NextAuthConfig;
