+++ src/modules/auth/types/auth.ts
@@ -6,8 +6,7 @@ export type Auth = {
 
 export enum ROUTES_AUTH {
     LOGIN = "/login",
-    SING_UP = "/signUp",
-    PRINCIPAL = "/principal",
-    HOME = "/home",
+    SING_UP = "/sign-up",
+    DASHBOARD = "/dashboard",
     RESET_PASSWORD = "/reset-password"
 }
